+++ tests/screenshot/apps/components/Chip.js
@@ -33,7 +33,10 @@ const chipSmokeTests = [
 	<Chip icon="home" deleteButton={{icon: 'closex', position: 'bottom'}} disabled>Disabled Chip with Icon and Delete Button (Bottom)</Chip>,
 
 	<Chip multiline>Lorem ipsum dolor sit amet, consectetur adipisicing elit. Amet at consequatur illum magnam omnis possimus, quas ratione rerum? Atque corporis magni nulla numquam odio officia quas quisquam vero! Repellat, vitae.</Chip>,
-	<Chip multiline icon="home">Lorem ipsum dolor sit amet, consectetur adipisicing elit. Amet at consequatur illum magnam omnis possimus, quas ratione rerum? Atque corporis magni nulla numquam odio officia quas quisquam vero! Repellat, vitae.</Chip>
+	<Chip multiline icon="home">Lorem ipsum dolor sit amet, consectetur adipisicing elit. Amet at consequatur illum magnam omnis possimus, quas ratione rerum? Atque corporis magni nulla numquam odio officia quas quisquam vero! Repellat, vitae.</Chip>,
+	<Chip multiline icon="home" deleteButton={{icon: 'closex', position: 'right'}}>Lorem ipsum dolor sit amet, consectetur adipisicing elit. Amet at consequatur illum magnam omnis possimus, quas ratione rerum? Atque corporis magni nulla numquam odio officia quas quisquam vero! Repellat, vitae.</Chip>,
+	<Chip multiline icon="home" deleteButton={{icon: 'closex', position: 'top'}}>Lorem ipsum dolor sit amet, consectetur adipisicing elit. Amet at consequatur illum magnam omnis possimus, quas ratione rerum? Atque corporis magni nulla numquam odio officia quas quisquam vero! Repellat, vitae.</Chip>,
+	<Chip multiline icon="home" deleteButton={{icon: 'closex', position: 'bottom'}}>Lorem ipsum dolor sit amet, consectetur adipisicing elit. Amet at consequatur illum magnam omnis possimus, quas ratione rerum? Atque corporis magni nulla numquam odio officia quas quisquam vero! Repellat, vitae.</Chip>
 ];
 
 const chipFocusTests = [
@@ -42,7 +45,8 @@ const chipFocusTests = [
 	<Chip icon="home">Focused Chip with Icon</Chip>,
 	<Chip checked>Checked and Focused Chip</Chip>,
 	<Chip disabled deleteButton={{icon: 'closex', position: 'right'}}>Disabled and Focused Chip</Chip>,
-	<Chip deleteButton={{icon: 'closex', position: 'right'}}>Focused Chip with Delete Button (Right)</Chip>
+	<Chip deleteButton={{icon: 'closex', position: 'right'}}>Focused Chip with Delete Button (Right)</Chip>,
+	<Chip multiline>Lorem ipsum dolor sit amet, consectetur adipisicing elit. Amet at consequatur illum magnam omnis possimus, quas ratione rerum? Atque corporis magni nulla numquam odio officia quas quisquam vero! Repellat.</Chip>
 ];
 
 const chipLargeTextTests = [
@@ -50,7 +54,8 @@ const chipLargeTextTests = [
 	<Chip>Default Chip</Chip>,
 	<Chip icon="home">Chip with Icon</Chip>,
 	<Chip checked>Checked Chip</Chip>,
-	<Chip deleteButton={{icon: 'closex', position: 'right'}}>Chip with Delete Button (Right)</Chip>
+	<Chip deleteButton={{icon: 'closex', position: 'right'}}>Chip with Delete Button (Right)</Chip>,
+	<Chip multiline>Lorem ipsum dolor sit amet, consectetur adipisicing elit. Amet at consequatur illum magnam omnis possimus, quas ratione rerum? Atque corporis nulla numquam odio officia quas quisquam vero! Repellat, vitae.</Chip>
 ];
 
 const chipTallglyphTests = [